requirements.txt
Flask==2.3.3
python-dotenv==1.0.1
gspread==5.12.0
oauth2client==4.1.3
Flask-Limiter==3.5.0
openai==1.35.10
httpx>=0.24,<0.26
supabase>=2.0.0
python-jose[cryptography]==3.3.0
